requirements.txt
numpy==1.18.5
matplotlib==3.3.3
opencv_python==4.5.2.54
torch==1.8.1
kornia==0.5.3
pandas==1.1.4
torchvision==0.9.1
kaggle==1.5.12
Pillow==8.2.0
